+++ src/errors.rs
@@ -19,8 +19,11 @@ pub enum FlokiError {
     #[error("Could not normalize the file path '{name}': {error:?}")]
     ProblemNormalizingFilePath { name: String, error: io::Error },
 
+    #[error("There was a problem rendering the template '{name}': {error:?}")]
+    ProblemRenderingTemplate { name: String, error: tera::Error },
+
     #[error("There was a problem opening the configuration file '{name}': {error:?}")]
-    ProblemOpeningConfigYaml { name: String, error: tera::Error },
+    ProblemOpeningConfigYaml { name: String, error: std::io::Error },
 
     #[error("There was a problem parsing the configuration file '{name}': {error:?}")]
     ProblemParsingConfigYaml {